// packages/zambdas/src/ehr/shared/pdf/types.ts
export interface ExternalLabsData {
  locationName: string;
  locationStreetAddress: string;
  locationCity: string;
  locationState: string;
  locationZip: string;
  locationPhone: string;
  locationFax: string;
  reqId: string;
  providerName: string;
  providerTitle: string;
  providerNPI: string;
  patientFirstName: string;
  patientMiddleName: string;
  patientLastName: string;
  patientSex: string;
  patientDOB: string;
  patientId: string;
  patientAddress: string;
  patientPhone: string;
  todayDate: string;
  orderDate: string;
  primaryInsuranceName?: string;
  primaryInsuranceAddress?: string;
  primaryInsuranceSubNum?: string;
  insuredName?: string;
  insuredAddress?: string;
  aoeAnswers: { question: string; answer: any }[];
  orderName: string;
  assessmentCode: string;
  assessmentName: string;
  orderPriority: string;
} // TODO: change this based on the actual data we need to send to submit-labs endpoint